Fix compiler warnings in test units
2 warnings in particular:
- one caused by subtracting LONG_MIN from LONG_MAX in FRANGE macro
  (fixed by casting to long double)
- one caused by usage of "%lf", which is a GNU extension (fixed by
  changing to "%f")

+++ test/src/test_struct.h
@@ -9,7 +9,7 @@
 
 #define LEN(X) (sizeof(X) / sizeof(*X))
 #define IRANGE(X,Y) ((X) + (rand() % ((Y) - (X) + 1)))
-#define FRANGE(X,Y) ((X) + ((double)rand() / RAND_MAX * ((Y) - (X))))
+#define FRANGE(X,Y) ((X) + ((double)rand() / RAND_MAX * ((long double)(Y) - (X))))
 
 struct data {
 	char    *name;
